+++ examples/palabra/structure
