+++ src/kathara_lab_checker/checks/CheckResult.py
@@ -1,6 +1,6 @@
 import logging
 
-from lab_checker.utils import green, red
+from kathara_lab_checker.utils import green, red
 
 
 class CheckResult:
