+++ src/kathara_lab_checker/checks/CollisionDomainCheck.py
