+++ src/kathara_lab_checker/checks/CustomCommandCheck.py
