+++ src/kathara_lab_checker/checks/ReachabilityCheck.py
@@ -2,7 +2,7 @@ import jc
 from Kathara.manager.Kathara import Kathara
 from Kathara.model.Lab import Lab
 
-from lab_checker.utils import get_output
+from kathara_lab_checker.utils import get_output
 from .AbstractCheck import AbstractCheck
 from .CheckResult import CheckResult
