+++ src/kathara_lab_checker/checks/protocols/AnnouncedNetworkCheck.py
@@ -3,9 +3,9 @@ import re
 from Kathara.manager.Kathara import Kathara
 from Kathara.model.Lab import Lab
 
-from lab_checker.checks.AbstractCheck import AbstractCheck
-from lab_checker.checks.CheckResult import CheckResult
-from lab_checker.utils import get_output
+from kathara_lab_checker.checks.AbstractCheck import AbstractCheck
+from kathara_lab_checker.checks.CheckResult import CheckResult
+from kathara_lab_checker.utils import get_output
 
 
 class AnnouncedNetworkCheck(AbstractCheck):
